pyproject.toml
[project]
name = "biasbuster"
version = "0.1.0"
description = "Toolkit for building curated training datasets to fine-tune LLMs for detecting bias in biomedical abstracts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx",
    "pandas",
    "tqdm",
    "anthropic",
    "nicegui>=3.8.0",
    "tokenizers>=0.22.2",
    "huggingface-hub>=0.34.0,<1.0",
    "argon2-cffi>=25.1.0",
    "bmlib[all]",
    "pdfplumber>=0.10",
]

[project.scripts]
biasbuster = "biasbuster.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
mlx = [
    "mlx-lm>=0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "attic/",
    "docs/",
    "eval_results/",
    "ollamazip/",
    "dataset/",
    "dataset_V2/",
    ".venv/",
    ".mypy_cache/",
    ".github/",
    ".claude/",
    "__pycache__/",
    "*.lora",
    "*.gguf",
]

[tool.hatch.build.targets.wheel]
packages = ["biasbuster"]

[tool.uv.sources]
bmlib = { path = "../bmlib", editable = true }

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
